split out DissociationConstants module
